requirements.txt
boto3==1.26.86
botocore==1.29.86
debugpy==1.6.7
django-debug-toolbar==4.1.0
django-csp==3.7
eventlet==0.33.3
gunicorn==21.2.0
jmespath==1.0.1
localstack-client==1.39
psycogreen==1.0.2
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
terraform-local==0.6
urllib3==1.26.14
XlsxWriter==3.1.2